Follow-up to 8373089853462e3e0eebb4e06131cb61c0dab2ab Reworked the transactional saving of PrusaSlicer.ini. Now in the configuration files is stored checksum to detect if the backup configuration isn't corrupted.
Reworked the error reporting, so that the user only gets an error on
reading PrusaSlicer.ini, not an error on handling the backup file.

Changed the format of the MD5 checksum line: Added "MD5 checksum "
to the exported line.

Fixed compilation for non-Windows platforms.
The opened backup configuration file is closed before removing it.